Update patch set 1
Patch Set 1:

Actually - it does not. There is no tradeoff/change. Anil pointed me at the fact that netty defaults are opposite to OS defaults (it does NODELAY by default).

So this change just "nails" the current behavior and documents the fact that this is what we expect now.
 
There is a more detailed discussion of this on its openflowforjava counterpart of this change. Basically, as long as we use netty (or similar pipeline processor) we have a double-buffering environment with netty (or similar) ensuring that one xmit request is ~ one transaction. This is the way we code for it by applying ser/des codecs to the pipeline. 

With this use pattern the NODELAY (netty current default) is the correct design. You do not want to buffer once more if you have buffered already. You want the whole transaction to be "committed" immediately.

So this is documenting current (desired) behavior just in case netty changes and/or someone replaces netty with something on a like-for-like basis.

Patch-set: 1