Update patch set 2
Patch Set 2:

For me, the choice to use a warning or debug log level depends on whether we consider the behaviour of the network equipments to be correct or not with respect to the NETCONF and XML standards. About that matter, I gave my opinion in a comment on the bug tracker: https://bugs.opendaylight.org/show_bug.cgi?id=2838#c6. Could you have a look at it and tell what you think about it?

Patch-set: 2