Update patch set 4
Patch Set 4:

(1 comment)

The refactoring is fine. Slight drawback is the additional object allocation for TransactionResources but it's negligible considering the rest of the transaction processing.

Perhaps the initialized flag and operationLimiter can also be added to TransactionResources so the TransactionFutureCallback can take a reference to TransactionResources instead of TransactionProxy. That would make it easier if we want to add specific unit tests for TransactionFutureCallback . But that could be looked into later.

Patch-set: 4
Label: Verified=0