Bug 3011: Changed scope of karaf:standard feature to runtime
Compile time scope of karaf standard feature caused
Karaf to install tomcat, jetty and lot of other features
via the startup.properties.

In Karaf 3.0.3 the tomcat feature works fine if installed
via the boot features in the features.cfg but if installed
via startup.properties, it prevents Karaf from starting.

Also added the eventadmin feature to the startupFeatures so it is
included in the startup.properties as the bundle does not work
properly for some reason if installed as a boot feature.

Change-Id: I309f7b116976493ba1377b442fb46f5a88f68361
Signed-off-by: Tony Tkacik <ttkacik@cisco.com>
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>