Update patch set 15
Patch Set 15:

The changes look good. A couple things re: compatibility...

Because the AppendEntries payload type has changed, this would affect rolling cluster upgrades somewhat as a Lithium-version leader wouldn't be able to lead a Helium-version follower. So de-ser would fail on the  Helium follower and no AppendEntriesReply would be received by the leader. The follower wouldn't receive any AppendEntries and would eventually trigger a re-election. This situation would resolve as soon as the follower is stopped for upgrade and I don't think there would be any long-term harm as raft is designed to be resilient. It may result in transaction failures during upgrade.

However, the AppendEntries ser/der was previously changed for Lithium making it incompatible from Lithium to Helium anyway. So rolling upgrades won't be smooth anyway.

The ModificationPayload class can actually be removed completely as it was introduced in Lithium (we can clean that up in a follow-up). The ShardTest#testRecovery method should be changed to test the new DataTreeCandidatePayload class. testHelium2VersionRecovery is already there for backwards compatibility.

We need a unit test for DataTreeCandidatePayload see/der.

Patch-set: 15
Label: Verified=0