CDS: rework components to support non-remote operations
Introduces ShardedTransactionChain (instead of TransactionChainProxy),
which tracks potential local transaction chains.

Updated TransactionProxy to hold a map of TransactionComponents -- each
of which acts as proxy encapsulating how to interact with a particular
shard. There are two basic flavors of components -- Local and Remote.
Local transaction components are further diversified to
read-onl/write-only/read-write kind.

Selection of appropriate component is done via
TransactionComponentFactory, of which we have two kinds -- Simple for
disconnected transactions and ShardedTransactionChain, which handles
waiting for previous transaction's proxies.

Change-Id: If5e09335827d58a74ba704a255d359290e508e5f
Signed-off-by: Robert Varga <rovarga@cisco.com>