WIP: CDS: rework components to support non-remote operations
Introduces ShardedTransactionChain (instead of TransactionChainProxy),
which tracks potential local transaction chains.

Introduced ShardedTransaction (instead of TransactionProxy), which
performs tracking of TransactionComponents, which act as proxies to the
sub-transaction backend. There are two favors of those -- remote ones,
which behave just as TransactionFutureCallback did and local ones, which
are instantiated only after we have ascertained that the shard is local
to us.

Change-Id: If5e09335827d58a74ba704a255d359290e508e5f
Signed-off-by: Robert Varga <rovarga@cisco.com>