Bug 3072: Make NeutronRoutersNorthbound.addRouterInterface() idem-potent
Neutron port used by the router interface is populated with
the deviceId and deviceOwner values before router interface add
message is received. The existing code for processing router
interface add did not get enabled until recently [1]. Now that
it does get called, it exposes this bug; where the checking
for deviceId and deviceOwner causes ResourceConflictException()
as it was not accounting for existing and new values being
the same.

Patch 2: Rephrase comment

[1]: https://review.openstack.org/#/c/174600/

Change-Id: I68d045bad7ee892f08610521f7f2788de147f63c
Signed-off-by: Flavio Fernandes <ffernand@redhat.com>