Allow customization of Raft Election mechanism
Sometimes we want to be able to choose exactly where a shard leader for
a given shard ends up. For example I might want inventory-config on member-1
and inventory-operational on member-2. Because Raft elections are random it
favors nodes that come up first and so the chances of all shard leaders
ending up on one node in the cluster is very high. This can hamper testing
scenarios where the leaders of these shards are distributed.

This patch provides a way in which a Raft election could be customized using a
strategy. The strategy has only one question can a actor become a leader or not.
The DefaultStrategy answers yes to that question.

Change-Id: Ie2d1963406d087f347d752529b18b8cfeef0584d
Signed-off-by: Moiz Raja <moraja@cisco.com>