Update patch set 6
Patch Set 6:

In looking at this some more, I'm wondering if the yang notification as defined will suit the needs for cluster-aware apps. Do apps really care about shard role changes in general? Specifically, do they care if the role changes from Candidate to Follower? That would expose raft details and that may not be a good thing (what if we change from raft to something else?).  Do they really care about shards? Hard-wiring app code to specific shards and shard names may be fragile. Shards are configurable in module-shards so a shard name could be changed or a module could be moved to another shard which might break code. Also in the future we intend to do sub-sharding and dynamic shards where data might move.

Apps identify their data in terms of yang paths so it seems to me any clustering API should be designed in terms of that. I think an app might want to ask "is my controller instance the leader for this yang path"? And "notify me when my controller instance becomes the leader for this yang path and when it is no longer the leader". In this way, the API talks "app speak" and we avoid exposing details about the shards themselves and in which shard a yang path resides.

In any event, I think we should discuss this with the community and get input from other folks. I don't want us to publish an API in a vacuum and get stuck having to support it for a couple releases if it turns out it doesn't fulfill the needs and/or is replaced by something else.

Patch-set: 6
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0