Update patch set 6
Patch Set 6:

Yeah - we would have to distinguish between oper and config - so data store type and yang path. I was just throwing something out there at a high level. 

The key point is that this needs more thought and we need to engage the community and app devs on what they require and what will work to make their apps work in a cluster.

Patch-set: 6