Update patch set 2
Patch Set 2:

As per my comment, I think this could allow a leader with an older log to get elected which goes against raft. 

I'm wary of putting in such a knob. I know Moiz intends this for testing but it will expose a cfg file that anyone could use if they get wind of it.

Patch-set: 2