Update patch set 2
Patch Set 2:

That's true if it was configured that way from the start. But if the cluster ran for a while then a follower that is now disallowed could have been the leader before.

I know that's probably contrived and this is intended for testing and the cfg file wouldn't be exposed by default so this hidden mechanism wouldn't be generally known. Still word can get around of this backdoor mechanism to force leadership that people might try to use not knowing that it's intended for testing. If the leader is forced to a specific node and that node is down, the cluster will be unable to elect a new leader and thus lose HA. 

I'm just playing devil's advocate. It's unlikely someone would misuse this knob.

Patch-set: 2