Update patch set 6
Patch Set 6:

Yes - opendaylight-inventory:nodes with sub-sharding would likely be a special case. Depending on how it's implemented, it could just be a logical aggregating shard with no actual leader. How would that work with this proposed API? Seems like no role change notifications would ever be sent. 

With an API based on yang IIDs, the answer to "is my instance the leader for opendaylight-inventory:nodes" might return unknown or "not applicable" or sort of lie and say yes. Not sure it would make sense for an app to ask that. But it would certainly make sense to query about "opendaylight-inventory:nodes/openflow:1".

Re: leaking of the raft states, I don't really see it as problematic - mainly I don't see a use case for apps needing to know about Candidate/Follower concepts. If that's the case, why leak (and document) that info to apps if you don't need to...

Patch-set: 6