Update patch set 26
Patch Set 26:

I found an issue with tx chain consistency that causes the unit test failure.

Sequence:

      DOMStoreWriteTransaction writeTx = txChain.newWriteOnlyTransaction();
       
      writeTx.write(TestModel.TEST_PATH, ...);

      final DOMStoreThreePhaseCommitCohort cohort1 = writeTx.ready();
 
      DOMStoreReadTransaction readTx = txChain.newReadOnlyTransaction();
      Optional<NormalizedNode<?, ?>> optional = readTx.read(TestModel.TEST_PATH).get(5, TimeUnit.SECONDS);
      assertEquals("isPresent", true, optional.isPresent());


Per the API contract, the readTx should see the write from the first writeTx but it may not due to timing. For the writeTx there's no knownLocal entry yet so it sends the FindPrimary message and creates a RemoteTransactionComponent, eliding the local tx path. Depending on timing, if the FindPrimary message returns before the read tx is created, the knownLocal map is updated and the read tx creates a local tx component. So we have a remote tx chain on the shard for the write tx and a local chain for the read tx. The read tx doesn't see the write b/c it exists on the  remote tx chain. This is can occur with cascaded writes as well - probably what caused the failure in my previous email. 

I think the most reliable way to make this work is to do the FindPrimary message synchronously.

Patch-set: 26
Label: Verified=0