Bug 3194: Dynamically update PrimaryShardInfo cache when leader changes
Added a PrimaryShardInfoFutureCache class which maintains the cache of
PrimaryShardInfo Future instances per shard and ShardInfoListener
registrations. This code was moved from the ActorContext.

The PrimaryShardInfoFutureCache instance is created by the
DistributedDataStore and passed to the ShardManager, as well as to the
ActorContext and TransactionContextFactory. The ActorContext uses
the cache as before and the TransactionContextFactory registers as a
ShardInfoListener, as it did before with the ActorContext.

On ShardLeaderChanged message, if the leaderId has changed, the
ShardManager calls remove on the PrimaryShardInfoFutureCache to
invalidate the entry for the shard name and nofify ShardInfoListeners.

Change-Id: I4c7a3845e311e5130d80f22f7cfb3c82c5ad731d
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>