Update patch set 1
Patch Set 1:

I know you fixed the generated equals but my point is that the CDS modules are not currently in a state where they can be dynamically restarted, at least reliably, due to issues in the CDS modules. So a restart of CDS may leave it in a bad state. Thus we don't want to allow it to be dynamically restarted at this point.  

So if someone did try to re-configure a CDS module via restconf, the controller would need to be restarted. However I don't see a use case for someone needing to do that anyway. An end user should not muck with the code dependency wiring - it is set out-of-box the way it needs to be. The config/operational properties can be set via a cfg file in etc (and it's dynamic).

Patch-set: 1