Update patch set 4
Patch Set 4:

This adds a dependency and assumption in the FeatureConfigPusher on the XmlFileStorageAdapter. It's possible to plugin in a separate persister adapter (not that anyone would actually do that :)). 

In my draft patch, I had made the Persister accessible from the ConfigPusher. We could add the getPersistedFeatures and setFeatureService methods to the Persister interface. This would decouple FeatureConfigPusher from the XmlFileStorageAdapter. What do you think?

So the feature pusher and persister pusher will continue to operate independent of one another. The feature pusher will push any config that isn't in the persisted config. I'm concerned this could result in an ordering issue, i.e. the feature pusher could push a config before the persister pusher has run. I think it might be better to have the feature pusher drive the whole process. It could first push the configs from the Persister before processing the installed features. That would make it at least somewhat deterministic.

Also, an upgrade from He wouldn't have the persisted features. So every restart will still have the potential clash  until a new config is pushed via restconf. I guess there is anything we can easily do about that.

Patch-set: 4
Label: Verified=0