Update patch set 4
Patch Set 4:

Yes, this adds dependency between FeatureConfigPusher -> XmlFileStorageAdapter. But I think this is the easiest way how to do this without changing the APIs and the control flow (there are 2 paths: 1: config-persister-impl that instantiates persister adapters and pushes initial config. 2: Feature-config-pusher that listens for feature events in karaf). If there are any other persisters, the behavior does not change, because the FeatureConfigPusher checkes whether the XmlFileStorageAdapter is in use. Changing the APIs of persister and config puhsher so that the FeatureConfigPusher can list the features would be a cleaner solution, but would require changing the APIs and would break any other config persister adapters. There are probably no other implementations, but its still changing the APIs before the release.

This commit is fully backwards compatible: it allows for other persister adapters that do not support karaf feature list, but also is capable of loading persisted configs without the feature list (in that case feature pusher pushes all the default configs from features = current behavior).

I think that the ordering has to be following:
- load current persisted config
- push current persisted config
- push default config from features, which are not persisted in the current config

And this is exactly what this commit does. Whether the feature list is retrieved from an API defined method or just a specific method for one persister (probably more suitable solution so close to release).

Patch-set: 4
Label: Verified=0