Update patch set 4
Patch Set 4:

I made some changes to simplify the code. I removed the ShardInfoListener and registration. The AbstractTransactionContextFactory updates the knownLocal map based on the PrimaryShardInfo returned from FindPrimary appropriately as it will only get transitions from local (DataTree present) to remote (DataTree absent) and vice versa. 

By removing ShardInfoListener, the TransactionContextFactory no longer needs the PrimaryShardInfoFutureCache nor does it need to keep track of the childChains.

From the unit tests, I found a bug where we need to differentiate between read-only, read-write, and write-only local chained transactions (you were right Moiz). I had taken a shortcut to just create a ReadWriteTransaction to encompass all. But this fails if you try to create a tx after a read-only tx in the chain b/c, by using a read-write tx under the hood for the read-only, it expects the previous read-write tx to be readied. So I had to add newReadOnly and newWriteOnly to LocalTransactionFactory which bubbled down.

Patch-set: 4
Label: Verified=0