Update patch set 4
Patch Set 4:

Patch Set 3 works OK for me.

However, I proposed an updated patch set which:
* Handles the corner case where the decoder receives only whitespaces without actual XML data (test case added)
* Adds a test case that contains all the white spaces we want to skip

At that point, the patch set does basically the same thing as my initial patch set, with the addition of improved logs (we log the skipped bytes) and unit tests.

Now, I'm still uncomfortable with the WARNING log level of the log messages issued when one or more leading whitespace is skipped. I have 3 objections:
1. As I already mentioned, for me this is not abnormal to have some line breaks between NETCONF messages (see https://bugs.opendaylight.org/show_bug.cgi?id=2838#c6 for a discussion of that point of view)
2. All the physical network devices I know send a line break between messages, and this generates a warning for each message. So in the end the logs contain a lot of such warnings.
3. What is an ODL operator or administrator supposed to do to fix the warning? Contact the vendor to change the behaviour of the device? As I believe the behaviour of the equipment is not incorrect and even desirable, I think this is not the way to go.

About (2), Robert Varga suggested to issue a warning only the "first time" a leading whitespace is encountered. It could be easily done on a per-NETCONF session basis with an extra member variable in the NetconfXMLToMessageDecoder class, but there would still be lots of warnings when the controller repeatdly establishes successive sessions with the same device.

About (3), the operator/administrator can silence the warning, but what's the point of the warning if the only possible action is to silence it?

So my suggestion would be to change the log level to TRACE.

Patch-set: 4