Make Election Timeout Factor configurable
Keeping the Heartbeat interval small and increasing the timeout
factor allows us to skip a few heartbeats if neccessary without
causing Followers to become Candidates.

Simply increasing the heartbeat interval has the negative effect
of causing replication messages for out-of date followers to go
slower so making the election factor configurable will be helpful
in keeping the cluster state stable.

To prevent spurious elections due to GC and thread latencies, the
default election timeout factor was increased to 20 as it is in Lithium.

Change-Id: I8b48401de58959b1c475f71945f89928d8b2a3cb
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>