Update patch set 6
Patch Set 6:

One note: This commit is related to the double config push for modules in features and in current, not the seamless upgrade from helium. So I think this discussion does not belong in here. But here are some thoughts:

I am not sure how we can guarantee the configuration from current to be pushed after all the required features are processed/pushed by the config feature pusher.

The feature pusher would have to know somehow when the last necessary feature was pushed and then allow the current pusher to push its current config. But the modified current config may contain modules that are linked to any feature (netconf-connector, bgp etc.). This information (list of required features) is missing in the feature pusher and there is no clear end of installing features. So the feature pusher cannot trigger push of current configuration. It simply does not know when to do so.

There could be a fixed delay for the current config push .. but thats hardly a good solution.

I can think of just one solution to this: Feature pusher would trigger the current pusher after all necessary features are pushed. And there are 2 ways for providing the list of necessary features: 1. Manually by user via a configuration option(in a config file in etc/ or a simple -D switch) or 2. Feature pusher scans the current config snapshot and finds all the namespaces there. Then it would try to satisfy each namespace by matching a namespace to a feature. After all namespaces are satisfied, current will be pushed. Feature pusher would have to scan every bundle of every feature to try find yang files in order to find the namespaces + related feature. Or we would have to provide a pre-defined map of namespace -> feature for quick lookup (but keeping such a map would be difficult).

Patch-set: 6