Bug 2594 -  PUT method returns wrong status for create resource
* Note: ietf-netconf-restconf-04 says:
"Consistent with [RFC 7231], if the PUT request creates a new resource,
a "201 Created" status-line is returned. If an existing resource is
modified, either "200 OK" or "204 No Content" are returned."

So we have to check it is exist or not.

TODOs - response obj. for PUT/MERGE BrocerFacade has to be redesign in berilium

Change-Id: I294e20a225f20e4a6114da00346b6058f91d99a3
Signed-off-by: Vaclav Demcak <vdemcak@cisco.com>