Update patch set 6
Patch Set 6: Code-Review+2

Yeah - this patch does do as advertised - getting an upgrade path to work can be done in another patch. I agree we can merge this one and refine it as needed to support upgrade.

This would be more manageable if we could know when all current features were processed after reboot. But karaf doesn't provide us a way to know that.

I wonder if we could do this via timestamps. So if the timestamp of a feature config xml file is newer than the current xml file, then push the one from the feature. On upgrade, if the files in etc/opendaylight/karaf were deleted prior to restart, then the features would extract the files with a new timestamp on re-install which would then be greater than the current xml file timestamp. However this would entail parsing the current xml file and reconstituting the original configs from the "blob". If that's possible then, once all the previously persisted features were processed, we would push the remaining configs in the current xml.

Patch-set: 6
Label: Code-Review=+2