Streamline updating out-of-sync follower
The first AppendEntries message a leader sends to a follower contains the
leader's entry at current commit index as it doesn't know the follower's next
index yet. If the tyhe previous index isn't present in the follower's
log, it sends back an unsuccessful reply. The leader then decrements the
prior index it just sent and sends the next message. This continue's
until the next index gets down to the follower's last index (or -1 if the
follower's log is empty). Then it succeeds on the follower and the leader
increments the next index and sends the rest.

This results in at least twice the number of messages sent to sync he
follower. The follower sends back its lastIndex in the unsuccessful
responses so it seems to make sense for the leader to use that index
instead of decrementing what it thinks might be the next index. This
eliminates the excessive AppendEntries messages.

Change-Id: Id00994dcffc287c227b2b4999bcbc21522b81d5f
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>