Update patch set 1
Patch Set 1:

We have tests other than the upstream integration tests that use the cars model. While they can be changed and semantically it may read better as "cars", I don't see a compelling reason to change it. It's just a test model... and others may be using it too.

Patch-set: 1