Update patch set 1
Patch Set 1:

Hmm... I am having some trouble overriding it externally (using -D) on the default Debian openjdk-7. Setting things in the global security.properties works, however it is not taking overrides despite the .override in the global file being set to true. 

As far as MINA - that is normal. Nearly all ssh implementations have the habit of doing their own crypto and ignoring global crypto settings. So whatever we do, we have to expect to handle ssh separately.

Patch-set: 1