Update patch set 1
Patch Set 1:

Well, it the user decides to override the defaults in their JRE installation and insists on insecure defaults -- we can't do much about that. The global solution works with default JRE installations, so we should be fine with that.

As for mina, I am fine with a point solution there. On second thought, I think it uses bouncycastle and maybe they have a global option, too ...

Patch-set: 1