Update patch set 1
Patch Set 1:

Bit curious when this situation would arise? 

For instance, this is for handling add-router-interface, it appears to say that if the port already has a deviceOwner of "router" or "gateway", then that's ok, otherwise it's an error, as it's assigned to a non-router.

When would a port be added to a router-interface and gateway at the same time ? They are two different neutron commands.

Patch-set: 1