Update patch set 1
Patch Set 1:

Yes the option is hardcoded, but it does not mean that users cannot change/disable it. The file can be deleted without any issues and JVM will simply ignore the option or it can be completely disabled in the primary java.security file.

This additional config does not override the primary file, it simply adds one more setting to it.

Users can provide specific settings in their primary file and disable the odl one from the same place.

So while I agree that using an env variable looks better, I am not sure if it gains us anything in this particular case.

Patch-set: 1
Label: Verified=0