BUG 2970 : Recovery fails with SchemaValidationException when removing modules
My prior fixes for bug 2970 did not address the recovery problem when a module is
removed but it's data is still present in the persisted files. This patch attempts
to address that.

I have added unit tests for ShardRecoveryCoordinator one of which fails because
pruning does not happen for the snapshot data. This is because it looks like when
you write data with "/" as the identifier the SchemaValidation is not kicking in
and the write which is expected to throw the SchemaValidationFailedException is
not doing so. This appears to be a bug in the datatree.

Change-Id: Id2793330441606c62dfd903cdd698c2f86b14c1b
Signed-off-by: Moiz Raja <moraja@cisco.com>