Update patch set 3
Patch Set 3: -Code-Review

Thats correct, after a restart the feature service does not provide complete features and this fix would remove a workaround for that.

So I think we cannot fix this issue, since the feature service does not seem to return consistent values. The only possible fix on our side would be trying to query the service in a loop with timeouts when it fails, which is far from an ideal fix.

Patch-set: 3