Bug 2948: Backport from Lithium
Backported "Log entries may be missing or not applied on persistence
recovery after prior snapshot": https://git.opendaylight.org/gerrit/#/c/18113 and
https://git.opendaylight.org/gerrit/#/c/18114

To get code reuse between the RaftActor and AbstractLeader wrt to
initiating a snapshot capture, I introduce a SnapshotSupport class.
This is a scaled down version of the SnapshotManager in Lithium. I
looked at backporting the SnapshotManager but it would've meant
backporting a significant amount of other refactoring so I opted
for the simpler approach.

Change-Id: Iaa434fbf562455adcc8b6a403b1250cb248cfbd4
Signed-off-by: Tom Pantelis <tpanteli@brocade.com>